renamed packages, fixes #11
