+++ src/main/java/it/gov/innovazione/lodviewng/LodviewNgApplication.java
@@ -1,4 +1,4 @@
-package it.innovaway.lodviewng;
+package it.gov.innovazione.lodviewng;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
