+++ src/main/java/it/gov/innovazione/lodviewng/controllers/ResourceController.java
@@ -1,11 +1,11 @@
-package it.innovaway.lodviewng.lodview.controllers;
+package it.gov.innovazione.lodviewng.controllers;
 
-import it.innovaway.lodviewng.lodview.bean.OntologyBean;
-import it.innovaway.lodviewng.lodview.bean.ResultBean;
-import it.innovaway.lodviewng.lodview.bean.TripleBean;
-import it.innovaway.lodviewng.lodview.builder.ResourceBuilder;
-import it.innovaway.lodviewng.lodview.conf.ConfigurationBean;
-import it.innovaway.lodviewng.lodview.utils.Misc;
+import it.gov.innovazione.lodviewng.bean.OntologyBean;
+import it.gov.innovazione.lodviewng.bean.ResultBean;
+import it.gov.innovazione.lodviewng.bean.TripleBean;
+import it.gov.innovazione.lodviewng.builder.ResourceBuilder;
+import it.gov.innovazione.lodviewng.conf.ConfigurationBean;
+import it.gov.innovazione.lodviewng.utils.Misc;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.StringUtils;
