+++ src/test/java/it/gov/innovazione/lodviewng/LodviewNgApplicationTests.java
@@ -1,4 +1,4 @@
-package it.innovaway.lodviewng;
+package it.gov.innovazione.lodviewng;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
